handles fractions of a sec for exposure times
[git-p4: depot-paths = "//depot/ACG/Smith/Firmware/": change = 95538]

+++ C++/include/PrintEngine.h
@@ -41,7 +41,7 @@ public:
     int GetPulseTimerFD();
     int GetExposureTimerFD();
     int GetMotorTimeoutTimerFD();
-    void StartExposureTimer(int seconds);
+    void StartExposureTimer(double seconds);
     void ClearExposureTimer();
     void StartMotorTimeoutTimer(int seconds);
     void ClearMotorTimeoutTimer();
@@ -50,7 +50,7 @@ public:
     void SendMotorCommand(unsigned char command);
     void Begin();
     void CancelPrint();
-    int GetExposureTimeSec();
+    double GetExposureTimeSec();
     int GetRemainingExposureTimeSec();
     void StopMotor();
     bool DoorIsOpen();
